+++ src/dev/messages/classNameIsDeprecated.js
@@ -1,0 +1,5 @@
+import console from '../Console'
+
+console.warn(
+  'The <b>className</b> attribute is deprecated. Use <b>class</b> instead.'
+)