+++ CHANGELOG.md
@@ -5,6 +5,10 @@
   API constructor `Node`, supertype of `HTMLElement`.
 * Removed the `HTTPRenderer`. Use the `StringRenderer` with whatever server instead.
 * The Hooks API has been overhauled from the Snabbdom one.
+* Snabbdom is configured with the `attributes` module instead of the `props` module, which
+  means that DOM API setters like `readOnly={true}` is not available. Use
+  `readonly={true}` like in HTML instead. `innerHTML` is still supported, and `className`
+  still works but is deprecated (see below).
 
 **Changes:**
 * `babel-polyfill` is no longer a peer dependency. Polyfill how you want.
@@ -13,6 +17,9 @@
   any version included in the application package.
 * `VirtualNode` instances expose a customized API on top of the Snabbdom interface. The
   TypeScript interface hides the Snabbdom interface as it is considered internal/private.
+
+**Deprecations:**
+* The `className` attribute is deprecated. Use `class` instead.
 
 # 0.4.0